fix: unable to access project (VF-000) (#5176)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements VF-000**

### Brief description. What is this change?

Fixed a major error with project sharing/import.
Right now when we are importing a project, we are looking if they have access to the original workspace, which is often no. (these projects like the ones on https://www.voiceflow.com/templates are in a different private workspace, and the project is "public"). Instead we should only be looking at if they have access to the workspace it is importing INTO.

![Screen Shot 2022-03-10 at 12 19 57 AM](https://user-images.githubusercontent.com/5643574/157594843-e17cd9bd-2043-4c72-bcf5-e5bca7c46001.png)

Co-authored-by: Tyler Han <tylerhan97@gmail.com>