feat: remove distinct platform (VF-3038) (#5190)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements VF-3038**

### Brief description. What is this change?
We have this odd pattern on the frontend, that was a legacy of a time when we wanted to upload to multiple platforms with the same project (publish to google & alexa on the same project, and use different intents depending on which platform you are on). 
You can see this frontend data format here:
![Screen Shot 2022-03-10 at 9 27 20 PM](https://user-images.githubusercontent.com/5643574/157798122-b1c874a2-c3ac-4e19-8180-be55b0f15960.png)
After this fix:
![Screen Shot 2022-03-10 at 10 53 21 PM](https://user-images.githubusercontent.com/5643574/157798897-49ebdf01-3371-436b-bf97-0b6a2254650f.png)

So interaction (choice), intent, and command blocks, always have this dictionary of { [platform]: data } within the data itself, but we only ever use the active platform's data.
The backend database has long abandoned this format. We are actually creating and removing it in the adapter layer.



Co-authored-by: Tyler Han <tylerhan97@gmail.com>