fix: socket connection and data loss (VF-000) (#5218)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements VF-000**

### Brief description. What is this change?

This might just be the biggest PR ever. Read my comments for more context.

> PLEASE TEST THIS OUT ON YOUR OWN AS WELL

## SocketIO Update
Since we moved from socket.io 2.x to 3.x, some [core changes needed](https://socket.io/docs/v3/migrating-from-2-x-to-3-0/#the-socket-instance-will-no-longer-forward-the-events-emitted-by-its-manager) (thank you @effervescentia) -`reconnect` events are no longer in the same namespace as custom events and instead need to be called through `socket.io` instead of `socket`. I've divided them up into separate enums. 
![Screen Shot 2022-03-18 at 12 29 49 AM](https://user-images.githubusercontent.com/5643574/158937717-8b720e07-b29e-4f1e-b43d-7be9a9675499.png)
![Screen Shot 2022-03-18 at 12 30 54 AM](https://user-images.githubusercontent.com/5643574/158937841-a3bde5a7-b04f-4bdc-93f1-c610ac607c45.png)

## State Issue
So I realized the biggest factors for desync/data loss: in `socket/client.tsx` if the `status` is anything except `SocketStatus.CONNECTED`, none of the diagram updates will send:
https://github.com/voiceflow/creator-app/blob/850726ca9faad5d12aa4cca2b67978a165f229c5/packages/creator-app/src/client/socket/diagram.ts#L34-L40
In the original code, the ONLY place we were calling `reestablishConnection` was from the `SocketEvent.CONNECT_ERROR` (`watchForConnectionError`) event.

The other big deal is that in `creator-api` we ABSOLUTELY RELY on the `socket.skillId`/`socket.diagramId` properties being set and joining the right channels:
https://github.com/voiceflow/creator-api/blob/master/lib/services/socketManager.ts#L72-L73

I've tested and confirmed if an instance of `creator-api` goes down, it only triggers a `SocketEvent.CONNECT_ERROR` event sometimes, and `SocketIOEvent.RECONNECT` in other events. It does not consistently classify switching servers as a "connection error". I think this depends on the load balancer/nginx layer.

So naturally, when an instance goes down for an update, the new server never gets called with `setupActiveDiagramConnection` and `socket.skillId`/`socket.diagramId` is never set (these properties are local to the server's active socket connection, and can't be transmitted through `pub/sub`.

## The fix
instead of doing `client.once(SocketEvent.INITIALIZE, ...` when there is a `CONNECT_ERROR`, I've added a dedicated listener so that whenever `INITIALIZE` happens, we always connect the diagram and set `status = SocketStatus.CONNECTED`. In the backend, it will only send `INITIALIZE` after the FE first sends it `INITIALIZE`. Which only happens on reconnection or page load - so it is fine.

### Prior Behavior
So given this logic, I would conclude when someone is on the canvas during a `creator-api` restart (not just updates, but any reason at all for a restart), if an additional person joins anytime during that person's session, there is a good chance their data would be overwriting each other because socket events are not being broadcasted anymore.

### Conclusion
Also this code is really hard to navgiate there is a lot of state manipulated in different places and sync issues between the redux realtime and the actual socket state. It will be ver nice when we get to remove this code.

Co-authored-by: Tyler Han <tylerhan97@gmail.com>