feat: implement UploadV2 component (VF-3355) (#5401)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements VF-3355**

### Brief description. What is this change?
Implement new components `UploadV2.Image`, `UploadV2.Audio` and `UploadBase`.

UploadBase implements all the underlying structure for upload components, and accepts a renderValue props, to render different kind of files. This component is not intended to be used alone. The idea is to create a component for each kind of file, and we use UploadBase to offer the base structure for them.

Also, simplified the structure for buttons, and made it work with namespace. So, instead of using PrimaryButtonLabel, now we can use `PrimaryButton.Label` for example.



Co-authored-by: vf-service-account <serviceaccount@voiceflow.com>