fix: select click bug (VF-000) (#5417)
context:

me and alex had added this prop because mike had noticed opening one dropdown in the project create modal wouldnt dismiss other dropdowns opened in the modal.

but alex had been testing via clicking the dropdown icon on the right, not the actual input itself. so now it has introduced this weird double click issue with the dropdown.

ive removed the prop which has fixed the double click issue, but the autodismiss issue is back (which is a much smaller issue)

the reason for this autodismiss issue iscause we stopPropagation the input click event, which prevents other dropdowns from knowing when to close.

i spent a good 30mins trying to figure out some logic in the select component to fix this, but to no avail as of yet.