feat: speed up prototype (VF-3534) (#5452)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements VF-3534**

### Brief description. What is this change?
TESTING ENVIRONMENT: https://creator-test-prototype.development.voiceflow.com/

As many of you know, Voiceflow is a "compiled" language, where `diagrams` are turned into `programs`. 
While this has advantages on the runtime side, it does mean anytime a change is made anywhere, we have to "recompile" all diagrams within a project, which results in a bit of a wait whenever a prototype is started.

What we can do is be a bit smarter about the compile process itself. Both practically and visually.
On `general-service` I worked with @MrPicklePinosaur to make these improvements:
* batch read of diagrams into the compiler, batch read from MongoDB (find)
* batch write of programs coming out of the compiler, batch write into MongoDB (bulkTransaction)
* the compile process used to be a full-on job within the job queue - this was a lot of overhead when the compile process, even for the largest project, doesn't take more than a few seconds. It is now **a single HTTP call** between `creator-app` and `general-service`

if a project had `N` diagrams, we were making **2N + 3** API calls before, between `general-service` and `creator-api`. After the `service` changes it effectively only does **3**.

On the creator side I've removed the loading circle for the prototype tool entirely - instead what happens, is that this chat bubble becomes the actual loading. The conversation won't kick off unless the compile job is done:
![Screen Shot 2022-05-18 at 8 38 25 PM](https://user-images.githubusercontent.com/5643574/169178024-72e72abc-6225-4168-9f41-20b9cf424c68.png)

This is just a sneaky way we can leverage some of the natural animations such as pressing the start button into time that's actually compiling in the background.

The goal is in the future
 

Co-authored-by: Tyler Han <tylerhan@Tylers-MacBook-Pro.local>
Co-authored-by: Tyler Han <tylerhan@Tylers-MBP.hitronhub.home>