fix: prune extra IF dynamic ports (VF-3597) (#5474)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements VF-3597**

### Brief description. What is this change?
I'm not sure how the original condition happened, but in the project described by the ticket, the user had 4 expressions in the IF step, but 5 dynamic ports. So there was clearly a desync somehow, but I'm not able to recreate it.

On the project, it doesn't accurately show up because https://github.com/voiceflow/creator-app/blob/f6c1b0c7bd7324ab6d13a5429eba508a4a847cb7/packages/creator-app/src/hooks/syncedLookup.ts#L3-L7 would always render an empty array.

The easiest way to deal with this is to `slice` the ports so that there can never be more ports than expressions. 
The only danger is we slice off valuable info - which I don't think we should be doing?

Co-authored-by: Tyler Han <tylerhan97@gmail.com>