feat: node port remap (VF-3521) (#5560)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements VF-3521**

### Brief description. What is this change?
We're going to narrow down 5 operations that affect the order of steps within a block:
`insert` `transplant` `isolate` `reorder` `removeMany`
Whenever we change the order we have to update the ports of the internal steps.

Examples:

- If I remove step#3 from a block, now all of step#2's ports has to point to step#4.
- If I remove step#5 and there is no step#6, all of step#4's ports should be pointing to null.
- If I add step#A1 step#A2 between step#B1 and step#B2, step#B1 needs to point to step#A1 and step#A2 needs to point to step#B2

Otherwise this will break the prototype experience because we rely on these hidden connections within a block.

`addBlock` adds an independent block + step - this doesn't affect anything.
`appendStep` seems like it only applies to command steps (which don't have ports).

For each operation I've injected a new property, `nodePortRemaps` - if there is a better name happy to implement it.
It looks like this:
```
interface NodePortRemap {
  nodeID: string;
  ports: {
    key?: string;
    type?: string;
    portID: string;
  }[];
  targetNodeID: string | null;
}
```
We just need to know the `nodeID`, all the ports of the `nodeID` that we are mapping, and what we are mapping them to.
Each of the 5 operations have special cases whether or not we should or should not be mapping ports. 

It was getting mad that I have more than 5 parameters for functions, so I changed it from `func(a,b,c)` to `func({ a, b, c })` in many places.

Co-authored-by: Tyler Han <tylerhan97@gmail.com>