fix: do not update if shallow partial equals (VF-3887) (#5650)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements VF-3887**

### Brief description. What is this change?
We do an absolute ton of `onBlur` save operations in the editors.
For example:
[Text Editor](https://github.com/voiceflow/creator-app/blob/master/packages/creator-app/src/pages/Canvas/components/TextListItem/index.tsx#L57)
[Speak Editor](https://github.com/voiceflow/creator-app/blob/master/packages/creator-app/src/pages/Canvas/components/SpeakAudioItem/index.tsx#L82)
[If Editor](https://github.com/voiceflow/creator-app/blob/master/packages/creator-app/src/pages/Canvas/managers/IfV2/components/ConditionsSection.tsx#L27-L29)
[Code Editor](https://github.com/voiceflow/creator-app/blob/master/packages/creator-app/src/pages/Canvas/managers/Code/CodeEditor.tsx#L53)
and many many more. 

The problem is that we send out a `Realtime.node.updateDataMany` action every time someone clicks on an input field and clicks off (focusing and blurring) even if nothing is changed. All these trivial requests get registered in the undo/redo history and have a lot of overhead for syncing between clients.

Because `onChange` is always sending partial data, the solution was to do a shallow partial comparison and not update if the shallow properties aren't changed. While it seems expensive to do a comparison, it saves a lot more overhead of not dispatching a redux action and subsequent logux syncs.

I tried my best to go through every editor and find all cases where blur might trigger a change.

also replace `shallowequal` with `Utils.object.shallowEquals` - its the exact same thing

Co-authored-by: Tyler Han <tylerhan97@gmail.com>