feat: implement new DesignMenu container (CT-17) (#5674)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements CT-17**

### Brief description. What is this change?
New container for Layers menu

About the new Resizable component.

The code to support both pixels and percentage and resize  for panels, was becoming too much complicated. 
I tried to simplified it, but I didn't get a nice result at the end.
So, I decided to create a new and tight component, to handle only single container resizes. It's pretty straight forward.

I end up implementing using classes, to leverage the use to class internal state and is useful in this kind of manipulation and hopefully in the future we can think about something similar for the original Resizable component, which by the way, we can call ResizableGroup, or something like that, to differentiate.