feat: fix ContextMenuContext api isOpen property (CT-456) (#5683)
Practically, you should now be able to change the color of a block that it positioned low enough on the window that the context menu popper opens on top of the block.

<img width="684" alt="Screen Shot 2022-06-28 at 12 56 43" src="https://user-images.githubusercontent.com/9748039/176225661-fb2f9872-471b-4d41-8a11-525013941d7a.png">


**About the PR**

This fixes, as a side-effect, the ContextMenuContext api. The `isOpen` property was always true. Now this is fixed and being used to prevent unintended `engine.clearActivation` calls.


Co-authored-by: Filipe Merker <filipe.merker@voiceflow.com>