fix: text markup desync (VF-3978) (#5702)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements VF-3978**

### Brief description. What is this change?
I was seeing this error every time I went back to the dashboard for certain projects: 
![Screen Shot 2022-06-30 at 5 11 20 PM](https://user-images.githubusercontent.com/5643574/176779009-53e89d03-d943-4acd-a113-9f0ba5c0e4a3.png)
and in the redux it is sending this action: note how everything is null already.
![Screen Shot 2022-06-30 at 5 11 32 PM](https://user-images.githubusercontent.com/5643574/176779054-18972d00-c366-4dae-b0e7-63c45742d320.png)

It turns out we have a debounced `useTeardown` in the `MarkupTextNode` component. So on component unmounting, we were waiting another 300ms and then trying to fetch the current metadata, when we are already on the dashboard.

I looked through the code carefully - we don't need this update to be debounced at all, it is only applied `onBlur`, this is actually a worse realtime experience because other users will experience 300ms of latency before the update.

We don't need to update on teardown either. If we refresh or close the window, onTeardown does nothing and the action won't be sent. If we go to another diagram or another window, the `onBlur` is automatically triggered before unmount anyways. I've tested this.
It's also important that we don't need to send a crapton of `node.UPDATE_DATA_MANY` requests every time we unmount, this bogs down the server.


Co-authored-by: Tyler Han <tylerhan97@gmail.com>