fix: infinite refresh (CT-543) (#5735)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements CT-543**

### Brief description. What is this change?
This took awhile to debug but this was SO annoying that I had to do it.

What is intended - during the lazy loading of a component, if the lazy load fails, either 
1. the component itself has some jank code causing it to crash (likely on development)  OR
2 the file it is referencing DNE anymore (like on prod, during a new deploy all the referenced chunks change names)

We want to force reload the window for the 2nd case so that the browser can pick up a new, up-to-date index.html file with the new references to chunks. However in the 1st case, we WANT it to crash so that we can read the error message.

# CURRENT BEHAVIOR

What is happening with the current `lazy()` code is on error, we add an `ALREADY_FORCE_REFRESH_KEY` flag to `sessionStorage` so if the page reloads and STILL crashes we just throw the error and read the mssage.
HOWEVER, we delete this `sessionStorage` flag whenever ANY component loads successfully: 
```ts
      const component = await factory();
      sessionStorage.removeItem(ALREADY_FORCE_REFRESH_KEY);
      return component;
```
This is what is causing that insane infinite reload issue we see on dev environments, and potentially on production. Here is what is happening, imagine we have 2 lazy load components: `Component1` and `Component2`

1. `Component1` loads correctly, `Component2` has some jank code error and crashes
2. this is caught by the `try/catch` in `lazy`, we check if the `ALREADY_FORCE_REFRESH_KEY` exists or not, it doesn't, so we add the flag and reload the window
3. `Component1` loads correctly AGAIN, but it also DELETES the `ALREADY_FORCE_REFRESH_KEY`
4. `Component2` has some jank code error and crashes, repeats step 2
5. infinite reload destroying your CPU and RAM :(

You can test this behavior very easily by adding this jank code to `pages/Project/components/Diagram.ts`:
```ts
// eslint-disable-next-line no-console
console.log((React as any).dne.dne);
```

# FIX
we need to tie each refresh token to the actual component that is being lazy loaded. Because `Component2`'s error triggering a reload or not has nothing to do with `Component1` loading correctly. Each lazy loaded component needs to maintain it's own unique token.

tbh the best way I've found to do this is just to hash [Function.prototype.toString](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Function/toString). I know it's kinda unsavory to do this, but there's no other good way get something unique in `lazy` (the components don't always load in the same order either, so counters ain't gonna work).

`Function.prototype.toString` works on all major browsers and guarantees uniqueness on built chunks or dev-hot-reload files. The actual lazy functions are pretty short too, they remain consistent even after refresh, which is important :)
```js
// original lazy load call
const Dashboard = lazy(() => import('@/pages/Dashboard'));

// factory.toString() on dev
"() => import('/src/pages/Dashboard/index.tsx')"

// factory.toString() on built
'()=>rf(()=>import("./index.3f17366c.js"),["assets/index.3f17366c.js","assets/RedirectWithSearch.0739147e.js"])'
```

because it is on `sessionStorage` the actual saved tokens are pretty inconsequential - they get deleted as soon as you close the tab so there's no fear of zombie ones or anything.

Co-authored-by: Tyler Han <tylerhan97@gmail.com>