fix: delete templates workspace (VF-4047) (#5740)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements VF-4047**

### Brief description. What is this change?
Deleting the the template workspace (team). This is not to be confused with the templates collection in mongo.

This was a really old feature where every new user would be added to a global workspace full of templates for them to copy.

We need to first delete all the workspaces, then delete all the functional code using it, then delete the column itself to not break anything. So the merge needs to be sequenced!

Co-authored-by: Tyler Han <tylerhan97@gmail.com>