fix: realtime markup overlay glitch (VF-4045) (#5773)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements VF-4045**

### Brief description. What is this change?
So this `resizeOverlay` function calls `engine.transformation.resizeOverlay(rect);` which actually draws the transform overlay over a markup node. 

Our translate api calls this `resizeOverlay`. 
https://github.com/voiceflow/creator-app/blob/f200ea987e2c0df2fd6a7296ca7de28b5fa5ff16/packages/creator-app/src/pages/Canvas/components/MarkupNode/hooks.ts#L127

On multiplayer this breaks because when someone else is dragging a Markup node, this `resizeOverlay` gets called on the node THEY are translating, and teleports your overlay to them.


Co-authored-by: Tyler Han <tylerhan97@gmail.com>