feat: models structure PART 1 (VF-4137) (#5839)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

# DO. NOT. MERGE. 
until infra has set up the required environment variables.

![Screen Shot 2022-08-01 at 7 14 50 PM](https://user-images.githubusercontent.com/5643574/182260747-ba73e8f1-0c3f-45a4-992a-a8eeccca519e.png)


**Fixes or implements VF-4137**

For local testing, simply add `.env.local` with `MONGO_URI` and `MONGO_DB` pointing to your `vfcli` environment.
Run with `yarn --cwd packages/realtime local`

### Brief description. What is this change?
We're going to change our architecture so that the realtime service directly reads and writes to the mongoDb. This is absolutely required for the `logux` library to prevent race conditions.
This is an experimental call that shows calling `mongoDb` in sequence, synchronously, ensures correct order of operations:
[test.zip](https://github.com/voiceflow/creator-app/files/9237773/test.zip)

To this end, I have added 2 new environment variables to `realtime`:  `MONGO_URI` and `MONGO_DB`.
I tried my absolute best to mirror the structure in `creator-api` because we're really just going to be copy and pasting code.
Instead of calling an API endpoint we're just going to skip the HTTP stuff and hit the model directly.

I added `models` as something for the service to build on initialization. Also ensured `clients` have a `setup` and `destroy` cycle too. This required some updates on `ml-gateway` as well.

# BREAKDOWN

this is only PART 1 of my fix, the ONLY thing it actually calls mongo for is fetching the diagram. I will address this in the comments. 
 

Co-authored-by: Tyler Han <tylerhan97@gmail.com>