fix: realtime node port fix (VF-4137) (#5844)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements VF-4137**

### Brief description. What is this change?
Hazah! I have finally figured out what is going on.
We actually never save `in` ports in the DB format - it is just assumed any node only has 1 in port, and when we load in the diagram through the adapter, we just take the `step`'s ID and add `"-in"` to the end.

When we create new steps through add block, or insert step, we are just generating a random ObjectID that never gets saved to the DB. So when you reconnect the ports just kind of disappear and the existing links are pointed to the wrong place.

`nodeDescriptorFactory` only produced nodes with `id: null` - I just added a new parameter so it takes in a nodeID as well.

Co-authored-by: Tyler Han <tylerhan97@gmail.com>