fix: aggressive socket timeouts (CT-000) (#5957)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements CT-000**

### Brief description. What is this change?
I did some testing of other frameworks. Our timeout is INSANELY high (20 seconds) - it means the user could be disconnected for up to 20 seconds, accumulating actions, and then the additional 10 second reconnect grace period we provide.

That means if I physically turn off my wifi, there is a 20 second timeout on the pong for the ping which is FAR too high. 
We would easily toss out an HTTP request if it was that bad. 

Other tools seem to instantly show it, here is me experimenting with a 1 second timeout:
<video src="https://user-images.githubusercontent.com/5643574/186238541-53526c79-0703-4674-9159-46fb2a3c7557.mov" />

I'm dailing it down to 5s but we can go lower than that.

This is what prod looks like today, with a crazy 20 second timeout:

<video src="https://user-images.githubusercontent.com/5643574/186248107-933fc11b-b449-41fd-af12-61d9f9423aba.mov" />




Co-authored-by: Tyler Han <tylerhan97@gmail.com>