fix: strip portsV2 from data adapter [bugfix] (CT-000) (#5997)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements CT-000**

### Brief description. What is this change?
This was quite a crazy bug that took me a long time to understand and reproduce. I found it by looking at our `realtime` datadog logs.

To reproduce:
1. make a text step
2. go to a different topic and come back
3. copy and paste the text step, then edit the contents of the new copied text step
4. go to different topic and come back
5. ports on new copied step fucked up

Video of reproduction:
<video src="https://user-images.githubusercontent.com/5643574/187598543-b8e52d53-c7f5-4127-b3ab-81d16e55988f.mov" />

What is happening is the `textAdapter` just returns the entire `node.data`:
https://github.com/voiceflow/creator-app/blob/bc4925f5e3ece85f830339239c375f3e33997a9a/packages/realtime-sdk/src/adapters/creator/block/base/text.ts#L6-L9
This `data` includes `portsV2`, which means it will enter our redux store as part of the node's data, ontop of the node's generic attributes (type, ports, children, etc.).
When a user edits a node's data in the editor, we call the `Realtime.node.updateDataMany` property and overwrite the `portsV2` on the DB.

When we copy and paste a step, we replace the portIDs with newly generated port IDs, but not the ports in the data. 
This resulted in duplicate ports: 2 blocks pointing to 1 port, and if one block gets deleted, we now have a port that has no parent.

This was a crazy error that I was only able to catch because of datadog and logrocket - the user never reported it.

Co-authored-by: Tyler Han <tylerhan97@gmail.com>