fix: prototype selectors (CT-000) (#6040)
### Brief description. What is this change?
These are small issues I noticed when testing prototypes and I want to enforce consistency on our prototype UI.


1. Topic Templates. Today, when we create a new topic it will be pre-populated with an intent event chip that says "Topic {X}" as the name.  However the name is actually the block name and there is no way to change it. Even after the user selects an intent, it will still show up as "Topix {X}". 
This also causes issues for Go To block and variable state's "Starting Block" selection - people can select the "Topic {X}" intent chip but none of the other event chips. This is not a good inconsistency to have. I think of chips and "blocks" as distinct entities.
My solution is just to populate new topics with an empty intent chip. There shouldn't be a special class of intent chip only on topic creation.
<img width="693" alt="Screen Shot 2022-09-11 at 7 36 46 PM" src="https://user-images.githubusercontent.com/5643574/189563974-fe6568a2-23eb-44e1-beb0-6a9901d380c4.png">
(From a code side, we use the block name on the chip because of the start step, I would like to avoid further conditions and stuff and keep the state simpler)

2. Chip Hover. When we hover a chip the icon switches to the "start test from here" play button. There are certain states where we shouldn't see the play button: when the prototype is already running, when dragging, etc. It turns out some of the original CSS logic might have been faulty and I have resorted to a more verbose solution.

Co-authored-by: Tyler Han <tylerhan97@gmail.com>