fix: prototype highlighting (#6039)
My attempt at fixing prototype highlighting, will be helped by this PR on general-service: https://github.com/voiceflow/general-service/pull/402

right now we have `activePathBlockIDs` and `activePathLinkIDs` that track highlighted steps across all diagrams we visit. We have a slight problem because the start block has the same ID on every diagram: `start00000000000000000000`

We have no idea of knowing if a start step on one diagram is highlighted or not, also we try to redraw every highlighted node for every diagram every time.

I've combined the two separate arrays into this format: `Record<string, { blockIDs: string[]; linkIDs: string[] }>` a dictionary defined by `diagramID`. This way we can confidently highlight the nodes for a particular diagram.

# engine issue
The other really really really big issue we have to is to do with engine. When we call `engine.prototype.start` it subscribes to the store and redraws everything for the current diagram.

However when we go into a separate diagram, we are calling `createEngine.clear()` but not tearing down or removing the old engine.
https://github.com/voiceflow/creator-app/blob/07d6fe78e6eb0471c9278f4fc6220ef5ee7bb76e/packages/creator-app/src/pages/Canvas/hooks/engine.ts#L47-L53

This means that `prototypeEngine`'s subscription is still active and the old engine is a memory leak.
My current solution is to just ALWAYS keep the same engine when switching between diagrams, and just rely on `engine.reset` to clear any intermediate state. We are not creating a new engine for every diagram, but keeping the same. I don't know if this is the best pattern.

We have a lot of old code that relies on the `EventualEngine` to become `null` then `engine`, however, there is ALWAYS an engine after opening the canvas now, so this old logic in `PrototypeTool/Trace` `await this.waitEngineAndNodes();` is totally wrong and did not wait fully. 




Co-authored-by: Tyler Han <tylerhan97@gmail.com>
Co-authored-by: vf-service-account <serviceaccount@voiceflow.com>