fix: realtime resource spikes (CT-000) (#6072)
Infra team has been reporting huge resource spikes on startup. We've narrowed down that it coincides with a rollout.
<img width="784" alt="Screen Shot 2022-09-15 at 6 31 43 PM" src="https://user-images.githubusercontent.com/5643574/190520046-89144d42-15dc-4732-9663-c171e14ae664.png">
![image](https://user-images.githubusercontent.com/5643574/190520086-22381d47-bf6d-4b7c-8165-ce0ea5686174.png)
![image](https://user-images.githubusercontent.com/5643574/190520094-c93c630d-c5c8-4b55-9fe1-b04802ddecbd.png)

This is mostly likely caused by users reconnecting to the logux server and then resubscribing. Because we only have 3 instances, if we have high load, this will cause a wave of resubscribe requests.

When someone loads the version channel, we do a huge amount of computation for the sake of showing steps from other diagrams (we load every diagram of a project into memory).

I think the short term solution is to just to not re-initialize everything on resubscribe and trust that the version channel is mostly synced. Again there is only really a 10 second window for people to make changes when disconnected - and it will be rectified on the next load. 

Co-authored-by: Tyler Han <tylerhan97@gmail.com>