fix: common job hook (CT-000) (#6073)
ExportContext and PublishContext are basically identical and we have a ton of code ensuring that jobs run properly and we poll their status. This should be easily reused.

Call it `start` instead of publish. We can also have a common jobclient pattern

I added a `startingOptionsCache` and `restart` calls so we can restart after finishing with the same settings

Co-authored-by: Tyler Han <tylerhan97@gmail.com>