feat: implementing custom blocks (PL-147) (#6028)
<!-- You can erase any parts of this template not applicable to your Pull Request. -->

**Fixes or implements PL-147**

### Important

The MVP is holding up 500K ARR in deals. I will try to address as much of the PR feedback as possible, however, due to time constraints (this must be done before the feature freeze) I will only guarantee that severe issues that set production on fire will get fixed immediately. 

**NOTE**: This feature will not be available to public. The feature flag is enabled only for particular customers that we're trying to seal deals with. 

### Brief description. What is this change?

Implements the Custom Block MVP feature, which allows users to create a "custom block", a user-defined block type which provides a designer-friendly interface around an existing block type (which can be a custom action for this MVP).

To create a custom block, you click "Add Custom Block" in the updated Library menu. 

<img width="348" alt="Screen Shot 2022-09-21 at 11 42 32 AM" src="https://user-images.githubusercontent.com/32404412/191549633-a7f38fc2-e27c-4532-bc93-8b045490d9f9.png">

This will bring up a Custom Block Editor, which allows you to define the custom block. Note that in the JSON editor, I have defined the variables, `{name}`, `{colour}`, `{height}`, and `{width}`. 

<img width="511" alt="Screen Shot 2022-09-21 at 11 43 47 AM" src="https://user-images.githubusercontent.com/32404412/191549941-56b8b514-8458-4a72-9b52-679e04c68e1a.png">

You can then drag-and-drop the newly created custom block onto canvas.

<img width="359" alt="Screen Shot 2022-09-21 at 11 45 33 AM" src="https://user-images.githubusercontent.com/32404412/191550365-aef613fd-021c-447c-b91e-ae2d0fd8740e.png">

<img width="472" alt="Screen Shot 2022-09-21 at 11 46 07 AM" src="https://user-images.githubusercontent.com/32404412/191550486-d77d8295-9f08-4abc-800a-aa4c9d6c6842.png">

When the user enters the canvas, they will see this custom side-editor interface. Notice how these labels match up with the variables I defined earlier. You can supply values for these variables.

<img width="458" alt="Screen Shot 2022-09-21 at 11 48 12 AM" src="https://user-images.githubusercontent.com/32404412/191550974-fc127b77-75a6-42b5-996d-0aaa26ebecc9.png">

In the prototype tool, the Custom Block behaves identically to the underlying custom action.

<img width="1055" alt="Screen Shot 2022-09-21 at 11 47 27 AM" src="https://user-images.githubusercontent.com/32404412/191550796-442220db-754a-4dc1-a6dd-974a4223b5b3.png">

Under the hood, the Custom Block is actually being compiled down to a Custom Action (see the related PRs section for the associated backend changes) and the values you provided for the variables will inserted into the JSON payload which is included in a Custom Action trace.