fix: transcript double messages (CT-1176) (#6266)
**Fixes or implements CT-1176**

### Brief description. What is this change?

An extra internal message was showing up as a user message when a user clicked on a button that triggers an intent. 
This message showed up because if it stored as a user intent request with no confidence but a matching intent. 

To prevent this from showing up as a user message, I implemented a check to see if the user request has an intent confidence or that the intent is not "None". This will allow no match responses to still show up but filter out these internal messages.